+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+### Added 
+
+- More tests
+
+### Changed
+
+- Newest Version of Auth Plugin
 
 ## [0.0.1] - 2019-09-27
 ### Added